+++ sub-packages/bionemo-scspeedtest/examples/scdataset_multiworker.py
@@ -13,12 +13,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+r"""This script benchmarks the performance of the scDataset wrapper for the SCDL and the AnnLoader with different numbers of workers.
 
-#!/usr/bin/env python3
-
-
-# Import AnnData support
-# from arrayloaders.io.dask_loader import DaskDataset
+Usage:
+    python scdataset_multiworker.py \
+        --adata-path /path/to/data.h5ad \
+        --scdl-path /path/to/scdl/
+The results will be saved to scdataset_multiworker_benchmark_{timestamp}_detailed_breakdown.csv
+"""
 
 import argparse
 from datetime import datetime
@@ -150,7 +152,7 @@ def create_scdl_scdataset_factory(
     return factory
 
 
-def comprehensive_benchmarking_example(
+def multiworker_example(
     num_epochs=1,
     num_runs=1,
     adata_path=None,
@@ -190,14 +192,10 @@ def comprehensive_benchmarking_example(
     print(f"All results will be saved to: scdataset_benchmark_{timestamp}_detailed_breakdown.csv")
     print()
 
-    # Parameters
-    # warmup_time_seconds = 30
-    # max_time_seconds = 120
-
     print(f"Benchmarking {num_runs} run(s) each")
     print()
     # =============================================================================
-    # Part1 2: SCDL Dataset with Multiple DataLoader Configurations
+    # Part 1: SCDL Dataset with Multiple DataLoader Configurations
     # =============================================================================
 
     # First run SCDL Regular as baseline
@@ -206,7 +204,7 @@ def comprehensive_benchmarking_example(
     for num_workers in num_workers_list:
         scdl_configurations.append(
             {
-                "name": f"Baseline SCDL_{num_workers}",
+                "name": f"Baseline SCDL_{num_workers}_workers",
                 "dataloader_factory": create_scdl_dataset_and_loader_factory(
                     batch_size=64, shuffle=True, data_path=scdl_path, num_workers=num_workers
                 ),
@@ -222,7 +220,7 @@ def comprehensive_benchmarking_example(
             for block_size in block_sizes:
                 scdl_configurations.append(
                     {
-                        "name": f"ScDataset_{block_size}_{fetch_factor}_{num_workers}",
+                        "name": f"ScDataset_{block_size}_block_size_{fetch_factor}_fetch_factor_{num_workers}_workers",
                         "dataloader_factory": create_scdl_scdataset_factory(
                             batch_size=64,
                             shuffle=True,
@@ -240,13 +238,15 @@ def comprehensive_benchmarking_example(
                     }
                 )
     # =============================================================================
-    # Part2: AnnData Dataset with ScDataset Configurations
+    # Part 2: AnnData Dataset with ScDataset Configurations
     # =============================================================================
     anndata_configurations = []
     for num_workers in num_workers_list:
+        if num_workers == 32:
+            continue
         anndata_configurations.append(
             {
-                "name": f"AnnLoader_Baseline_{num_workers}",
+                "name": f"AnnLoader_Baseline_{num_workers}_workers",
                 "dataloader_factory": create_annloader_factory(batch_size=64, shuffle=True, num_workers=num_workers),
                 "num_epochs": num_epochs,
                 "max_time_seconds": max_time_seconds,
@@ -260,7 +260,7 @@ def comprehensive_benchmarking_example(
             for block_size in block_sizes:
                 anndata_configurations.append(
                     {
-                        "name": f"ScDataset_AnnData_{block_size}_{fetch_factor}_{num_workers}",
+                        "name": f"ScDataset_AnnData_{block_size}_block_size_{fetch_factor}_fetch_factor_{num_workers}_workers",
                         "dataloader_factory": create_scdataset_annloader_factory(
                             batch_size=64,
                             shuffle=True,
@@ -279,16 +279,16 @@ def comprehensive_benchmarking_example(
     benchmark_dataloaders_with_configs(
         dataloader_configs=scdl_configurations,
         shared_dataset_factory=None,  # Each config creates its own dataset
-        output_prefix=f"scdataset_benchmark_{timestamp}",
+        output_prefix=f"scdataset_multiworker_benchmark_{timestamp}",
     )
     benchmark_dataloaders_with_configs(
         dataloader_configs=anndata_configurations,
         shared_dataset_factory=create_anndata_dataset_factory(adata_path),
-        output_prefix=f"scdataset_benchmark_{timestamp}",  # Same file as SCDL results
+        output_prefix=f"scdataset_multiworker_benchmark_{timestamp}",  # Same file as SCDL results
     )
 
     print("Benchmarking completed!")
-    print(f"All results saved to: scdataset_benchmark_{timestamp}_detailed_breakdown.csv")
+    print(f"All results saved to: scdataset_multiworker_benchmark_{timestamp}_detailed_breakdown.csv")
     print()
 
 
@@ -336,7 +336,7 @@ if __name__ == "__main__":
         "--num-workers",
         nargs="+",
         type=int,
-        default=[0, 1, 2, 4, 8, 16, 32],
+        default=[4, 8, 16, 32, 64],
         help="List of block sizes to test. Default: %(default)s",
     )
 
@@ -357,7 +357,7 @@ if __name__ == "__main__":
 
     print("BioNeMo Benchmarking Framework - ScDataset Test")
     print("=" * 80)
-    comprehensive_benchmarking_example(
+    multiworker_example(
         num_epochs=args.num_epochs,
         num_runs=args.num_runs,
         adata_path=args.adata_path,
